requirements.txt
langchain>=0.3.0
langgraph>=0.2.34
faiss-cpu>=1.8.0
pydantic>=2.8
pydantic-settings>=2.4
python-dotenv>=1.0
sentence-transformers>=3.0
transformers>=4.44
torch>=2.2
numpy>=1.26
scikit-learn>=1.5
rank-bm25>=0.2.2
unstructured>=0.14.6
pdfminer.six>=20231228
beautifulsoup4>=4.12
openai>=1.45
tiktoken>=0.7
pyyaml>=6.0.2
rich>=13.8